Rename css and sass filenames
